requirements.txt
numpy>=1.19.5
pandas>=1.2.4
streamlit>=0.82.0
scikit-learn>=0.23.1
matplotlib>=3.2.2
scipy>=1.5.0
xgboost>=1.4.1
pillow>=7.2.0                  
imagesize>=1.2.0